Change return type of fatorial method from int to long for better precision

+++ src/main/java/edu/puc/utils/Calculator.java
@@ -1,7 +1,7 @@
 package edu.puc.utils;
 
 public class Calculator {
-    public static int fatorial(int posicao){
+    public static long fatorial(int posicao){
         if (posicao <= 1){
             return 1;
         }
